Move wrapText to util
