+++ src/util/index.ts
@@ -28,3 +28,47 @@ export function convertToSentenceCase(string: string): string {
             return c.toUpperCase();
         });
 }
+
+export const wrapText = (
+    text: string,
+    maxCharPerLine: number
+): {
+    formattedText: string;
+    numLines: number;
+} => {
+    let numLines = 0;
+    const insertBreaks = (text: string): string => {
+        numLines++;
+        if (text.length <= maxCharPerLine) return text;
+
+        const words = text.split(" ");
+        if (words.length === 1) return text;
+
+        let currentLineLength = 0;
+        let numWordsInCurrentLine = 0;
+        for (let i = 0; i < words.length - 1; i++) {
+            // +1 to account for space between words
+            currentLineLength += words[i].length + 1;
+            numWordsInCurrentLine++;
+            if (currentLineLength + words[i + 1].length > maxCharPerLine) {
+                break;
+            }
+        }
+        const textInCurrentLine = words
+            .slice(0, numWordsInCurrentLine)
+            .join(" ");
+        if (words.length > numWordsInCurrentLine) {
+            return (
+                textInCurrentLine +
+                "<br>" +
+                insertBreaks(words.slice(numWordsInCurrentLine).join(" "))
+            );
+        } else {
+            return textInCurrentLine;
+        }
+    };
+    return {
+        formattedText: insertBreaks(text),
+        numLines: numLines,
+    };
+};
